Completed Plan and Unplan encounter Save
