+++ src/dataEntryApp/reducers/programEncounterReducer.js
@@ -1,0 +1,96 @@
+const prefix = "app/dataEntry/reducer/programEncounter/";
+
+export const types = {
+  //to get planned program encounters using enrolment uuid
+  GET_PROGRAM_ENROLMENT: `${prefix}GET_PROGRAM_ENROLMENT`,
+  SET_PROGRAM_ENROLMENT: `${prefix}SET_PROGRAM_ENROLMENT`,
+  //to get unpanned program encounter using program uuid
+  GET_UNPLAN_PROGRAM_ENCOUNTERS: `${prefix}GET_UNPLAN_PROGRAM_ENCOUNTERS`,
+  SET_UNPLAN_PROGRAM_ENCOUNTERS: `${prefix}SET_UNPLAN_PROGRAM_ENCOUNTERS`,
+
+  GET_PROGRAM_ENCOUNTER_FORM: `${prefix}GET_PROGRAM_ENCOUNTER_FORM`,
+  SET_PROGRAM_ENCOUNTER_FORM: `${prefix}SET_PROGRAM_ENCOUNTER_FORM`,
+  // GET_PROGRAM_ENCOUNTER: `${prefix}GET_PROGRAM_ENCOUNTER`,
+  SET_PROGRAM_ENCOUNTER: `${prefix}SET_PROGRAM_ENCOUNTER`,
+  SAVE_PROGRAM_ENCOUNTER: `${prefix}SAVE_PROGRAM_ENCOUNTER`,
+  UPDATE_OBS: `${prefix}UPDATE_OBS`
+};
+
+export const getUnplanProgramEncounters = (subjectTypeName, programUuid) => ({
+  type: types.GET_UNPLAN_PROGRAM_ENCOUNTERS,
+  subjectTypeName,
+  programUuid
+});
+
+export const setUnplanProgramEncounters = unplanProgramEncounters => ({
+  type: types.SET_UNPLAN_PROGRAM_ENCOUNTERS,
+  unplanProgramEncounters
+});
+
+export const getProgramEnrolment = enrolmentUuid => ({
+  type: types.GET_PROGRAM_ENROLMENT,
+  enrolmentUuid
+});
+
+export const setProgramEnrolment = programEnrolment => ({
+  type: types.SET_PROGRAM_ENROLMENT,
+  programEnrolment
+});
+
+export const getProgramEncounterForm = (encounterTypeUuid, enrolmentUuid) => ({
+  type: types.GET_PROGRAM_ENCOUNTER_FORM,
+  encounterTypeUuid,
+  enrolmentUuid
+});
+
+export const setProgramEncounterForm = programEncounterForm => ({
+  type: types.SET_PROGRAM_ENCOUNTER_FORM,
+  programEncounterForm
+});
+
+export const setProgramEncounter = programEncounter => ({
+  type: types.SET_PROGRAM_ENCOUNTER,
+  programEncounter
+});
+
+export const updateObs = (formElement, value) => ({
+  type: types.UPDATE_OBS,
+  formElement,
+  value
+});
+
+export const saveProgramEncounter = () => ({
+  type: types.SAVE_PROGRAM_ENCOUNTER
+});
+
+export default function(state = {}, action) {
+  switch (action.type) {
+    case types.SET_PROGRAM_ENROLMENT: {
+      return {
+        ...state,
+        programEnrolment: action.programEnrolment
+      };
+    }
+    case types.SET_UNPLAN_PROGRAM_ENCOUNTERS: {
+      return {
+        ...state,
+        unplanProgramEncounters: action.unplanProgramEncounters
+        //programEncounters: action.programEncounters
+      };
+    }
+    case types.SET_PROGRAM_ENCOUNTER_FORM: {
+      return {
+        ...state,
+        programEncounterForm: action.programEncounterForm
+      };
+    }
+    case types.SET_PROGRAM_ENCOUNTER: {
+      return {
+        ...state,
+        programEncounter: action.programEncounter
+      };
+    }
+    default:
+      return state;
+  }
+}
