+++ src/dataEntryApp/sagas/index.js
@@ -4,6 +4,7 @@ import { all, fork } from "redux-saga/effects";
 import subjectDashboardSaga from "./subjectDashboardSaga";
 import generalSubjectDashboardSaga from "./generalSubjectDashboardSaga";
 import programSubjectDashboardSaga from "./programSubjectDashboardSaga";
+import programEncounterSaga from "./programEncounterSaga";
 import { saveUserInfoWatcher } from "rootApp/saga";
 import programSaga from "./programSaga";
 // import translationApiSaga from "../sagas/TranslationSaga"
@@ -17,7 +18,8 @@ export default function* rootSaga() {
       generalSubjectDashboardSaga,
       programSubjectDashboardSaga,
       saveUserInfoWatcher,
-      programSaga
+      programSaga,
+      programEncounterSaga
       // translationApiSaga
     ].map(fork)
   );
