+++ src/dataEntryApp/views/subjectDashBoard/components/ProgramEncounterForm.js
@@ -2,7 +2,7 @@ import { connect } from "react-redux";
 import FormWizard from "dataEntryApp/views/registration/FormWizard";
 import { ObservationsHolder, Individual, SubjectType } from "avni-models";
 //import { updateObs, saveProgramEnrolment } from "dataEntryApp/reducers/programEnrolReducer";
-import { updateObs, saveProgramEncounter } from "dataEntryApp/reducers/programReducer";
+import { updateObs, saveProgramEncounter } from "dataEntryApp/reducers/programEncounterReducer";
 import { withRouter } from "react-router-dom";
 import { setValidationResults } from "dataEntryApp/reducers/registrationReducer";
 
@@ -10,20 +10,12 @@ let subject = Individual.createEmptyInstance();
 subject.subjectType = SubjectType.create("Individual");
 
 const mapFormStateToProps = state => ({
-  form: state.programs.programEncounterForm,
+  form: state.dataEntry.programEncounterReducer.programEncounterForm,
   subject: state.dataEntry.subjectProgram.subjectProgram,
-  // observations: [],
-  // obsHolder: new ObservationsHolder([]),
-  observations: state.programs.programEncounter && state.programs.programEncounter.observations,
-  obsHolder:
-    state.programs.programEncounter &&
-    (state.programs.programEncounter.observations
-      ? new ObservationsHolder(state.programs.programEncounter.observations)
-      : new ObservationsHolder([])),
-  //state.programs.programEncounter.observations
-
-  //   observations: state.dataEntry.enrolmentReducer.programEnrolment.observations,
-  //   obsHolder: new ObservationsHolder(state.dataEntry.enrolmentReducer.programEnrolment.observations),
+  observations: state.dataEntry.programEncounterReducer.programEncounter.observations,
+  obsHolder: new ObservationsHolder(
+    state.dataEntry.programEncounterReducer.programEncounter.observations
+  ),
   title: `New Encounter`,
   //   saved: state.dataEntry.enrolmentReducer.saved,
   //   onSaveGoto: "/app/subject?uuid=" + state.dataEntry.subjectProfile.subjectProfile.uuid,
